Experiment with caching the Develocity local cache
This can be useful when pushing small iterations to a pull request.

It is limited to pull requests only and only for the Initial JDK build.
We will check if it improves things and the size of the cache.

Related to #43559

(cherry picked from commit 287ce49c1366d4fc33721bb774702da74d2d2e6d)