Configure Maximum Message Size for Vert.x gRPC Server
Previously, the Vert.x gRPC server was created without passing any configuration. This commit creates an instance of `GrpcServerOptions` to configure the maximum message size for the server. At the moment, there is no other configuration, but it will be extended with Vert.x 5.

To maintain consistency, we reuse the maxInboundMessageSize property instead of introducing a dedicated one.

Fixes https://github.com/quarkusio/quarkus/issues/43573