Undertow - Only load resources that are known
If we try to load every resource, we end up creating cache entries in
the resourceCache CHM of the AppClassLoader, which is definitely not
what we want in this case, given we are supposed to serve only known
resources.

Now I must admit the fix looks a bit too simple to be true :).

Fixes #43676