Add Brotli4J Native Dependencies to the BOM
Quarkus automatically includes the Brotli native dependency corresponding to the platform on which the application is built. However, this approach doesn’t work when the application needs to support multiple platforms. In such cases, users must manually include the appropriate dependencies.

This commit adds Brotli4J native dependencies to the BOM and updates the documentation to guide users on which dependencies to add for various platforms.