Properly apply the update recipes in version order
This is extremely important as some recipes are additive and changing
the order will affect the result.
I noticed that because a project updated from 3.5 to 3.15 was updated to
use quarkus-resteasy-client-jackson instead of
quarkus-rest-client-jackson due to the recipes not applied in order.

(cherry picked from commit c211c89aa92bbffd2a797606d65008bf592c1ae2)