Ensure test-callbacks are removed when switching from `@QuarkusTest` to `@QuarkusIntegrationTest`
In case there are both test classes annotated with `@QuarkusTest` and `@QuarkusIntegrationTest`, JUnit callbacks (those in `AbstractTestWithCallbacksExtension`) are not properly cleaned up, causing the (later) run integration tests to try to call the "old" callbacks, which cannot work (class loader issue). This change adds a `Runnable` to clean up the memoized callbacks in `QuarkusTestExtensionState.close()`, which works for both `IntegrationTestExtensionState` and `QuarkusTestExtension.ExtensionState`.