REST Server/Client allows configuring the removal of trailing slashs
Before these changes, the REST server and client extensions were removing the trailing slashes that were set by users in the `@Path` annotations.

For example, when creating a resource like:

```java
@Path("/a/")
@Produces(MediaType.TEXT_PLAIN)
public class HelloResource {

        @GET
        public String echo() {
            // ...
        }
}
```

The effective path was `/a` instead of `/a/`. Note that it does not matter whether we place the `@Path` annotation at method level because the behaviour will be the same.

At the client side, when having the following client:

```
@RegisterRestClient(configKey = "test")
@Path("/a/") <1>
public interface HelloClient {
        @GET
        @Produces(MediaType.TEXT_PLAIN)
        @Path("/b/")  <2>
        String echo();
}
```

In this case, the trailing slash will be removed from <1> but not from <2>.

After these changes, we are adding the following properties:
- `quarkus.rest.removes-trailing-slash`

To configure the server extension to not remove any trailing slash from the `@Path` annotations.

- `quarkus.rest-client.removes-trailing-slash`

To configure the client extension to not remove any traling slash from the `@Path` annotations used at interface level. The annotations set at method level will behave the same.

Note that this does not introduce any change in behaviour, so everything should keep working as before unless users use the new properties.