Remove `cleanOnValidationError` config in Flyway
Starting in Flyway 11.0.0 (see https://documentation.red-gate.com/flyway/release-notes-and-older-versions/release-notes-for-flyway-engine), the `cleanOnValidationError` function and configuration has been removed.
An error will be thrown if this feature is configured.

Therefore it is better to remove this configuration from Quarkus so applications using it would fail as well (instead of deprecating it)