Reimplement @RegisterForReflection(registerFullHierarchy)
The idea is to leverage what we already have with the
ReflectiveHierarchyBuildItem infrastructure.

It simplifies things but I also did several changes for more
consistency:
- the @RegisterForReflection config is now passed to
  ReflectiveHierarchyBuildItem, which wasn't the case before
- we register the nested classes by default consistently with what
  @RegisterForReflection is doing by default

IMO, we need more consistency in this handling because we are using the
same annotation so the behavior should be consistent.
If you have one class registered for reflection and you decide to
register the hierarchy instead, the behavior should be the same.