OpenTelemetry: fix the Redis instrumenter in case of a tainted connection
The Vert.x Redis client does not always report the database instance number,
because sometimes, it simply isn't known. Specifically, that occurs when
the connection is tainted; for example, when a custom database is selected
using the `SELECT` command.

In this case, the trace simply will not include the dabatase instance number.