Register JPA static metamodel classes for reflection
So that Hibernate ORM can populate their class_ attribute.

Alternatively we could have prevented Hibernate ORM from populating the
class_ attribute, but this would have been a breaking change in JVM
mode or for applications that already enabled reflection on the static
metamodel. This is safer.

I'm not adding a test because another pending PR is going to tighten up
native compilation checks, and those would fail if this PR doesn't work,
so we'll know.