Redis Client: fix NPE when constructing XPendingSummary
Certain fields that were deemed to be mandatory can in fact be `null`,
as demonstrated in newly added tests.

Also, this commit fixes translating `XTrimArgs` to the actual argument
array in case `MAXLEN` is set to `0`. That is, in fact, a valid value,
as demonstrated by the newly added tests and also by inspecting Redis
source code [1].

[1] https://github.com/redis/redis/blob/7.2.4/src/t_stream.c#L941