Maven Resolver: initialize Maven properly, especially the local repository
The `BootstrapMavenContext.newRepositorySystemSession()` method is now
a lot closer to `DefaultRepositorySystemSessionFactory.newRepositorySession()`,
especially around how it initializes the `LocalRepositoryManager`. It is now
done _after_ setting up the config properties on the `RepositorySystemSession`,
which has also been fixed (to include the properties of all active profiles,
including those in `settings.xml`).

Furhter, this commit renames 2 files to match the name of the (non-`public`)
class declared in them, and fixes self type emulation in `BootstrapMavenContextConfig`.