Allow extensions to contribute datasouces
Making this build item public should allow extension to contribute extensions
at build time. This allows the possiblity of defining datasources as code rather
than config.

Runtime attributes still need to be provided as config, however it is already possible
for extensions to provide runtime config so this is not really a concern.