Migrate Reactive Pg Client extension
Deprecate PgPoolBuildItem for removal

Refactor PgPoolCreator producer and processor

The PgPoolCreator change will not break existing implementations.
Indeed, the contract now uses the parent Pool type, so returning PgPool is fine.

In the recorder, the pool configurator now returns the parent Pool type.
Then, in the processor, new bean configurators are set up to support injection of the different pools:

- bare pool
- vendor pool
- mutiny pool
- mutiny vendor pool

Also, we can get rid of vendor pool lookup in ReactivePgDataSourcesHealthCheck.
The PgPoolSupport bean indicates which datasource names correspond to a PgPool that has actually been created.
It allows to filter Pool bean instances in ReactivePgDataSourcesHealthCheck.