Introduce support for JEP 483's new AOT cache
This works in much the same way as AppCDS does -
it allows us to create an archive at build time
without the training run.
This is done by launching the application up to
a known point where no connections are made,
therefore capturing a good portion of the classes
that will be needed for the entire startup without
the hassle of introducing a proper training run.