Simplify and fix building the list of framework endpoints
We already have the full path at some point so there's really no need to
rebuild it once again, especially since the code was incorrect as it was
adding the /q/ prefix always if it wasn't there in the path.

It's perfectly valid to not have /q in the path, for instance if you set
quarkus.smallrye-health.root-path to /something/else.

I added a new test based on the report in addition to the ones we added
recently for related fixes.

Fixes #46040