Clarify expected behavior of SmallRye Config disambiguation.
Per discussion in #46245, previous versions of SmallRye Config
supported configuration disambiguation for optional config keys
with dynamic values. This was fixed in smallrye 3.10.0 and shipped
to Quarkus users in 3.17.0.Final. These documentation changes help
clarify expected behavior of Quarkus, how to disambiguate configuration,
and the limitations of the existing support.