Continuous Testing: add support for build system like test selection
This commit adds support for selecting tests to run during continuous testing
that is very close to the native test filtering capability of the given build
system. That is, in Maven, there's support for the `-Dtest=...` system property,
while in Gradle, there's support for the `--tests` command line option. Format
of these options follows the format used by the build system.