Logging with Panache: add support for method references of Log methods
In case the `Log` API is used in a lambda expression, everything already works,
because a lambda expression body is compiled into a separate method and
the `Log` API is invoked using classic `invokestatic` instruction.

However, in case the `Log` API is used as a method reference (`Log::info`),
it breaks, because the `LoggingWithPanacheProcessor` fails to transform that
usage. This commit fixes that by looking for `invokedynamic` invocations
of the `LambdaMetafactory` where the method handle belongs to `Log`. Such
instructions are rewritten to delegate to the JBoss Logging instance, as usual.