ArC: fix disposer resolution in case the disposed parameter declares no qualifiers
In case a disposed parameter of a disposer method declares no qualifiers,
the disposer method was considered matching even for a producer that does
declare qualifiers. This is wrong, because per the typesafe resolution
rules, if there's no qualifier, `@Default` is assumed.

This commit fixes that: in case no qualifier is declared, `@Default` is
assumed, so that the disposer only matches a producer that also declares
no qualifier (or explicitly declares `@Default`).

(cherry picked from commit 5ea026a1e5975b4f839215c995cd89dcca6f77bf)