Execution Model Validation: properly discover Dev UI JSON RPC classes
The `JsonRPCProvidersBuildItem` often used to be produced only in dev mode
(but note that some extensions produce it always already). This disallowed
using this build item to discover Dev UI JSON RPC classes properly, so
we used to ship a gross hack in the execution model annotation validation.

It is time to fix that. The `JsonRPCProvidersBuildItem` is always produced,
so we can rely on it for discovering all the Dev UI JSON RPC classes.