Jandex: reindex if the index is too new
We've been bitten by this issue many times in the past: if a dependency
upgrades to a newer Jandex than what Quarkus contains, build fails
with an `UnsupportedVersion` exception.

This commit is a proper solution: if the exception is thrown, we swallow it
and reindex the archive with the Jandex version that Quarkus incorporates.

Further, this commit increments the minimum Jandex version we "trust" from
Jandex 2.1 to Jandex 3.0. We rely on new features from Jandex 2.4 and 3.0
on many places, so it makes sense to make 3.0 the new minimum. The changes
added in 3.2 and 3.3 are also important, but are not used heavily yet.