ArC: @Singleton is no longer treated as bean defining annotation in strict mode
This is because `@Singleton` is not a bean defining annotation in CDI, full stop.
We want to treat `@Singleton` as bean defining in the default mode, just because
it makes sense, but not in strict mode, even though it is not tested in the TCK.

Some tests that run in the strict mode and expect `@Singleton` to be a bean
defining annotation are also adjusted in this commit (they use `@Dependent`
instead).