Allow adding Open API specification endpoint to OpenAPI using config
This mirrors the PR that allow for adding the health endpoint to
the OpenAPI config [1] to also do that for the documented Open API
endpoint itself.

Since the server dynamically generates this file, it can be requested
from the server. Thus, it should also be possible to specify it in
the document.

To avoid a recursive problem of needing to specify the full structure
of the document itself, inside of the document, we only specify that
it is a JSON object.

[1]: https://github.com/quarkusio/quarkus/pull/12044/