Disable checks related to unsupported bytecode enhancement
And switch back to the legacy behavior of "hoping for the best".

In Quarkus, we expect model classes to be enhanced.
Lack of enhancement could lead to many problems,
from bad performance, to Quarkus-specific optimizations causing errors/data loss,
to incorrect generated bytecode (references to non-existing methods).

So, in ORM 6.6 we've introduced checks to detect cases where bytecode
enhancement is not possible, so that developers can fix their apps.

We had a rough start, the checks were under/over-reporting,
leading to bug reports, so we fixed the checks.

We're now doing the best we can, and it turns out
the checks still have have many false positives --
in particular they assume property access for
mapped-superclasses and embeddables,
because the access type cannot be determined locally for those,
which leads to reporting some very valid setups as invalid.

We've tried our best, and it turns out the remedy is worse than the
disease. Let's disable these checks.

Longer-term, the solution is just to make bytecode enhancement work even in these cases:
https://hibernate.atlassian.net/browse/HHH-18825