Don't use logging on failed start
There is a good change that logging has not been setup yet, and the logging will not be written out.

A better approach would probable be to register a shutdown hook in the delayed log handler to make sure the messages are written out before exit, but this is better than nothing.