Add infrastructure to disable GIB and Develocity on specific PRs
By adding labels:
- `ci/disable-gib` to disable GIB
- `ci/disable-develocity-cache` to disable the Develocity build caching

It can be useful to test specific changes thoroughly.

Note that the labels have to be added directly when you create the pull
request or you will have to push again to the pull request after adding
the labels: rerunning the workflow won't be enough.

Please use responsibly.