SmallRye Fault Tolerance: skip private methods
Private methods can never be intercepted, so it makes no sense to check
if they are fault tolerance methods or to register them as such.

In fact, when compiling to native image, this used to cause a problem with
reflection registration. ArC automatically registers intercepted methods
for reflection, but it doesn't register `private` methods (because they
cannot be intercepted). SmallRye Fault Tolerance tries to reflect on all
fault tolerance methods, which used to fail on private methods. This commit
fixes that.