Restore previous behavior of QuarkusTransaction#isActive()
And mark is as deprecated.
It was causing issues with IronJacamar so we have a proof that this
change is disruptive. And transactions are critical enough that we don't
want to take any risks.

This is a follow-up of https://github.com/quarkusio/quarkus/pull/47212.

And related to
https://github.com/quarkiverse/quarkus-ironjacamar/issues/174.

(cherry picked from commit f974e2c5e46d08e2c4c8c205ad0baed2277b3df8)