Fully disable max-parallel to see if it causes Spring native issue
For whatever reason, the Spring native job is sometimes pending forever
and it seemed to happen quite a lot lately.

Let's see if it could be related to max-parallel being set (even if we
set it to 999 when using RunsOn so it shouldn't be in the way...).

(cherry picked from commit 050a11ce90e7988994cf19a62c6db931afd7e2f2)