Micrometer: remove MicrometerCounted
This annotation was likely created in the past because `@Counted` couldn't be
put on the interceptor class, because it was `@Target(METHOD)` [1]. Nowadays,
`@Counted` can be put on a class, so this workaround is no longer necessary.

The class is in the `runtime` package and isn't documented anywhere, so this
commit just removes it directly.

[1] That actually wasn't necessary. We could have created an annotation
transformation to "put" the annotation on the interceptor class, but for some
reason, we didn't. Doesn't matter anymore.