Better handle registry referencing non-existing codestarts
We need to be a lot more solid when the registry is referencing
non-existing codestarts.
Which can happen if an extension is pointing to a non-existing codestart.

The patch is extremely small but we have to push the logger there thus
why patch is larger than expected.

Related to
https://quarkusio.zulipchat.com/#narrow/channel/187030-users/topic/.60quarkus.20create.20cli.60.20is.20possibly.20broken.3F/with/517462042

(cherry picked from commit 3ee44a41abd8020a3ff3a5eb60391922342665f1)