Make placing JPA listener annotations a little safer
This essentially handles the case where such a listener
is added on private method of a CDI EntityListener class.
When the listener is a @Singleton we can safely transform
the private method to package private, but when it's
@ApplicationScoped there is nothing we can do
(as any transformation we apply is invisible to Arc
when it creates the Client Proxy class), so we
therefore fail the build (with an actionable error
message).

Closes: #24340