Retry when `CL.definePackage()` throws an IAE
`java.lang.ClassLoader.definePackage()` succeeeds if, and only if, the package was not already defined. This is problematic when packages are defined concurrently, for example when classes are loaded concurrently. `j.l.ClassLoader` does not offer something like `definePackageIfNotExists()`.

The proposed fix here is to change the implementations of `definePackage` in in `io.quarkus.bootstrap.classloading.QuarkusClassLoader#definePackage` and also `io.quarkus.launcher.RuntimeLaunchClassLoader#definePackage`, the same change was applied by #42139 to `io.quarkus.bootstrap.runner.RunnerClassLoader#definePackage`:
1. check if the package is already defined via `CL.getDefinedPackage()`
2. if it is, then return
2. if not, then call `CL.definePackage()`
3. if the former fails with an `IllegalArgumentException`, then check once against `getDefinedPackage()`, rethrow the IAE, if the pacakge's not defined

I could not find another `ClassLoader` implementation in Quarkus, which could be prone to IAEs via `CL.definePackage()`'s behavior.

Fixes #37363