Runtime initialize classes to preserve Netty's assumptions
Runtime initialize classes to allow netty to use the field offset for
testing if unsafe is available or not.

In java 24 and higher, unsafe is not available resulting in `UNSAFE`
being set to `null`.

In NettySubstitutions Quarkus substitutes the value of
`REFCNT_FIELD_OFFSET` which is done without the need of unsafe being
present. This results in `unsafeOffset` not returning -1 which netty
expects when unsafe is not available. As a result Quarkus native builds
with this substitution fail with unsafe is not available.

Closes: https://github.com/quarkusio/quarkus/issues/47903