Fix Platform overriding of codestart dockerfile data
We need the Platform to override the values in the codestart data, and
for that, they need to be prefixed by the codestart name.
TBH, my first inclination was to always override properties in the
codestarts with what comes from the data... but it breaks a test so
let's not go there for now.

Also fix the native-micro entry as it was missing a `.from`.

And make sure the data coming from previous platforms is properly fixed
when extracted.

Fixes #48102