Redis Client: fix BLPOP and BRPOP
The timeout value was added to the command twice, which is incorrect.
The correct position is _after_ the keys; inserting it before the keys
makes the timeout value a key itself (one that likely does not exist).

That doesn't need to be especially problematic with a standalone client,
because Redis will just ignore keys that do not exist, but it becomes
especially problematic with a cluster client, because the command may
suddenly target multiple nodes. This ends up with an error immediately,
on the client side.