ArC: remove Gizmo 1 FieldDescriptors and deprecate Gizmo 1 MethodDescriptors
These classes are no longer used in ArC, because all code was migrated
to Gizmo 2 and uses `{Field,Method}Descs` instead.

Hence, `FieldDescriptors` can be removed, because it isn't `public`.
`MethodDescriptors`, on the other hand, is `public`, so we cannot
remove it immediately. Instead, we deprecate it for removal.

Note that the Gizmo 2 `{Field,Method}Descs` are both not `public`,
so they cannot be used outside ArC.