Avoid copying mvnw in our sources
Follow-up of https://github.com/quarkusio/quarkus/pull/46525
and also related to https://github.com/quarkusio/quarkus/pull/48506

We need the Maven wrapper to be present in some modules to avoid having
the nested Maven runs inherit from the root .mvn/maven.config.
But it's better to copy it in the build when needed.

Per gripe from @aloubyansky.