Redis Client: improve handling of timeouts in list commands
Since Redis 6, all list commands that take a timeout accept a floating
point value in addition to an integer. To allow usage with old Redis
servers, we make sure that if the given timeout is a whole number of
seconds, the ultimate command sent to Redis uses an integer number.
Only if the given timeout is _not_ a whole number of seconds do we use
a floating point value.

(cherry picked from commit 78071279b5ee02c2845606218771b3c33ea9c828)