Attempt to reduce spurious linkage errors
Since the `RunnerClassLoader` is (more or less) lockless, if a class is concurrently redefined, we can get a spurious `LinkageError` which is consumed and ignored. We can reduce this incidence by doing one last probe for the loaded class before we attempt to define it but after we have done all of the resource loader work.