Handle `docker-compose` output more like we did before
Before, we redirected error to output and then logged the output. We do not want that now, because then we would not gather the error output on failure, but we can still log the same way we did before by capturing the error output and writing it to the log line-by-line.