Set priority for form auth handler
It used to be more or less randomly defined depending on the order of
execution of the build steps, which made it work until we started
playing with the order for build reproducibility.

The patch is large because I moved the priorities to the runtime module
so that I could use them in the recorder.