Make sure that we test context propagation between rest server and client
There's a change in Vert.x 4.5.14+ that stops copying context locals
from the parent to the child duplicated contexts, so we have to do it
via nested contexts (where we manually copy the locals).

This is required for things like propagating HR transactions/sessions.

From @gsmet: Apparently, 3.15 is not affected but it's nice to have a test.

Related to #49163 and #49133