First pass at moving our tests to recent Keycloak version
We are running a very old legacy version for testing which is a bad
idea.
Especially since this version is using a very old JDK.

This is a first pass, it is still very rough and some tests are not
passing yet but I would like some feedback from a full CI run.

Co-authored-by: Michal Vavřík <mvavrik@redhat.com>