Core: fix dumping debug class/source files
Previously, the given name was used to construct the dump file path.
The given name may be delimited using `.` or `/`, which leads to
a strange situation where some debug files are dumped into a directory
hierarchy (`com/example/foo/bar/MyClass.class`), while other debug files
are dumped into the root directory (`com.example.foo.bar.MyClass.class`).

With this commit, we no longer use the deprecated `getName()` method
to construct the dump file paths. Instead, we always use the `internalName()`
method, which returns a `/`-delimited name, so that all files are dumped
into a directory hierarchy. For log messages, on the other hand, we always
use the `binaryName()` method, which returns a `.`-delimited name, so that
logging uses classic Java class names.