Redis Client: fix *SCAN cursor handling
`*SCAN` cursors in Redis are strings that contain _unsigned_ 64-bit integers.
Until this commit, we ignored that and treated cursors as _signed_ integers,
which may lead to an exception when parsing the number. Further, this allowed
us to devise a sentinel value, `-1`, to mark the "not yet started" situation.

Since this commit, we parse and serialize cursors as unsigned integers. This
disallows using -1 as a special value, so we add a special flag `initial`
to mark the "not yet started" situation instead.

This commit also deprecates (for removal) the access to cursor values. This is
specifically:

- `Cursor.INITIAL_CURSOR_ID`
- `Cursor.cursorId()`
- `ReactiveCursor.INITIAL_CURSOR_ID`
- `ReactiveCursor.cursorId()`

The `INITIAL_CURSOR_ID` value used to be `-1`, which is incorrect. The initial
_and_ final value of a cursor is always `0`.

The `cursorId()` returns the cursor value as `long`, which makes it super easy
to treat it as a signed integer, which is wrong.

We would be able to add a `cursor()`/`cursorValue()`/`cursorString()` method
that returns the cursor value as a `String`, but I'm not sure there's any actual
need for access to the underlying value, so I'm refraining from this for now.