Improve initialization of constant pool scanning configuration
In BytecodeTransformerBuildItem, we can define a set of classes which
will be used to determine if a class is affected by a transformation by
by checking the constant pool.

This is most notably (and I think exclusively) used as an optimization
for the Panache transformation of public fields -> accessors.

When you have an application with a lot of entities, all the
BytecodeTransformerBuildItem targeting the classes using these entities
are carrying the Set of all the entities.

And we used copy the Set over and over for each class affected by the
transformation, which was generating a lot of useless allocations.

Obviously, this optimization only works if we don't use the constant
pool scanning optimization for other transformations.

And that why, I ended up adding another optimization: we now only pass
to the transformer the entities we have detected as being used in this
class instead of the whole list of entities.
While doing that should make the first optimization a lot less useful, I
think it's still good to keep it around.