Filter bytecode transformers on a case per case basis
We used to apply all transformers if at least one of them was not
requiring any constant pool check, including the ones for which we had a
constant pool check, which was a bad idea: we now filter the
transformers properly and we apply the ones that should be applied.

Nice bonus: we don't need to prepare some sort of aggregated config
before applying the transformers as we don't need one.

All in all, it's faster and should be more correct.