Use Set.of() instead of Collections.emptySet() in core parts
Marko told me that it was actually faster to iterate and from what I can
see it's true.
I won't change it everywhere but it seems like a good idea to make the
change in the Core parts.

Note that some of the changes I made might not be critical but it wasn't
worth the hassle to check everywhere if it actually made a noticeable
difference given the size of the change.