Reduce our dependency to exotic common-io versions in tests
This is problematic as soon as we are depending on commons-io in our
build.
We were fortunate enough for it to not have been a problem until now,
even if I'm pretty sure Maven makes use of it (but probably only very
common methods).
This is a problem now that we use commons-compress to build the jars.