Use JSON instead of Java serialization for ApplicationModel
This allows Develocity to auomatically replace the file paths in this model
when running tests on a remote machine, where the paths are different than
locally.

We use jackson, because it has the necessary flexibility to handle the application
model with its deep nesting and various subclasses without having to make many changes
to the model itself.