ArC: fix construction of interception proxies
During execution of the constructor of an intercepted non-contextual instance
(which is a superclass of an interception proxy, `*_InterceptionSubclass`),
virtual method calls were supposed to be forwarded to the delegate. This isn't
correct, because the `delegate` is only set at after the superclass constructor
finishes; therefore, all such attempts used to fail with `NullPointerException`.

This commit fixes that by adopting the same strategy we use for client proxies
and classic interception subclasses (for contextual instances): virtual calls
in the constructor are invoked on `this` (or, in fact, on `super`) and
forwarding is not attempted. Note that this has to be checked both for
intercepted and non-intercepted methods.