Reduce pinning by preloading superclasses
This works by reducing the amount of time spent by the JVM calling back into the class loader, which is generally all pinned time. Instead of having the JVM recursively load supertypes of the loaded type, we do it ourselves on the Java side ahead of time so when the JVM loads the superclass, it is already present.