Redis Client: improve and consolidate error handling in Redis transactions
The error handling is conceptually simple and uniform across
the blocking API and reactive API.

If the transaction block succeeds, the transaction is `EXEC`uted.
In that case, some commands may succeed and some may fail; this is
standard Redis behavior. The `TransactionResult.hasErrors()` method
indicates whether at least one command failed. The failures are
represented as `Throwable`s in the `TransactionResult` list.

If the transaction block fails, the transaction is discarded and
the exception is rethrown.

In case of the optimistic locking API, if the pre-transaction block
succeeds, the transaction is started and execution continutes
to the transaction block.

If the pre-transaction block fails, all watched keys are `UNWATCH`ed
and the exception is rethrown. In this case, the transaction is not
started and the transaction block is not executed.