Simplify inference of the SupportedDbKind
We used to try to derive it from the dbKind string, falling back to
attempting to interpret the Hibernate ORM "DB Product Name".

But the "DB Product Name" can only be set in situation where *we have a
dbKind string already*, so the fallback is pointless.

Since the fallback also has problems (use of CommunityDatabase which is
not necessarily in the classpath), we might as well remove it.

(cherry picked from commit 6eb9e8bbe9a7c1e99afe7c4a95242e21ab786aed)