PathTreeVisit - Use a very simple implementation to relativize Path
When in PathTreeVisit, we are in a very controlled environements, and we
shouldn't have bad surprises with Paths.
I stumbled upon the fact that relativizing the Paths was actually quite
costly in allocations, especially for UnixPath.
This PR introduces an extremely simple implementation that should be
good enough in the controlled environment that is PathTreeVisit.

I checked that all the relative Paths generated in my sample app were
equals to the old ones.