Reduce allocations due to ClassLoader#defineClassSourceLocation
It will call URL#toExternalForm() for each class, which will generate
again and again the same string pointing to the jar files.

By introducing a specific URLStreamHandler, we are able to avoid calling
URL#toExternalForm() and use a cached version of it for each
JarResource.

While not making a tremendous difference, it is an easy way to reduce
our allocations at startup.