Rework Hibernate tests around "no datasource" cases
The existing "NoConfigNoEntitiesTest" do not test that, they are strictly
equivalent to ConfigNoEntitiesDefaultPUDatasourceUrlMissingDynamicInjectionTest,
so they can be safely removed.

Testing there is no datasource requires that the Datasource extensions
not produce an implicit datasource, which would require that we don't
add a driver/client to the classpath. Since that's not practical, we'll
explicitly disable the implicit datasource instead.